backend/requirements.txt
flask
flask-cors
pandas
matplotlib
openpyxl
google-generativeai
